+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [1.7.0](https://github.com/tigrisdata/objgit/compare/v1.6.0...v1.7.0) (2026-09-23)
+
+### Features
+
+- add signed push webhooks and hook file metadata ([#13](https://github.com/tigrisdata/objgit/issues/13)) ([34e61cf](https://github.com/tigrisdata/objgit/commit/34e61cf7e253593b3848d1db56cae6ddc1d8b3f2))
+
+# [1.6.0](https://github.com/tigrisdata/objgit/compare/v1.5.1...v1.6.0) (2026-09-23)
+
+### Features
+
+- add kubernetes deployment, healthz endpoint, and image ci ([#14](https://github.com/tigrisdata/objgit/issues/14)) ([38e173f](https://github.com/tigrisdata/objgit/commit/38e173f443169a68b1c11c5f46259c04d87b8b2d))
+
 ## [1.5.1](https://github.com/tigrisdata/objgit/compare/v1.5.0...v1.5.1) (2026-09-11)
 
 ### Performance Improvements